Add FfiBindingError
This allows bindings (ex JNI) to pass errors received from a callback
back through libsignal-protocol-rust so they can be returned by the
operation which had invoked the callback.

We could instead do something fancy with Any here instead, so that a
callback into Java which throws an exception ends with exception being
captured, flowing through libsignal-protocol-rust and then rethrown
when we return to Java. But it's not clear if that's so necessary
vs the much simpler option of capturing the exception text.