Fix sender key logic
The Java storage interface wants to return a newly allocated
SenderKeyRecord if it is new. This seems weird to me vs just returning
null but in any case detect a fresh record.

We continue to expect an Option<> and create a new fresh record if
None is returned as other bindings may find it more convenient to just
return None/null/whatever.

Also limit sender key IDs to 31-bits for Java