Swift: Remove 'throws', using the Java API as a reference
While the FFI level of the Rust client is very conservative and can
fail for things like null pointers, Swift restores a lot of invariants
that C can't guarantee. In order to strike a balance between recovery
and ergonomic use of the API, this commit removes 'throws' from APIs
based on whether the corresponding Java API throws a checked
exception.

Note that these changes do not *precisely* match the Java API; there
are a few places where the Java API is declared as throwing because a
precondition was previously lazily enforced, but libsignal-client
enforces it eagerly. Example: PreKeySignalMessage.getIdentityKey (the
validity of the identity key is checked on deserialization).

Accessors that no longer throw have also been converted to properties.