FFI: drop build.rs that invokes cbindgen
Using cbindgen in a Cargo build script is convenient, but it means
cbindgen can't reuse the results of the actual build for macro
expansion, which results in doing a good chunk of a /second/ build.
Move that logic to swift/build_ffi.sh instead.

Depends on having the latest cbindgen.