Swift: fix Xcode build on Big Sur
Rust build scripts and proc-macros are compiled for the host, rather
than the target, which means they need a search path to link against
the host's libraries. macOS Big Sur doesn't provide that in the
default search paths anymore, so we add one manually based on an
environment variable provided by Xcode.