Update dependencies
This is `cargo update` along with manually upgrading jni to 0.19 and prost to 0.7

This isn't a complete cargo update as it avoids updating polyval as there is
a duplicate dependency issue there in cpuid-bool between the newest version of
polyval crate and the current latest version of the sha2 crate.

The bytes dependency in protocol that was causing duplicate dependencies was a
red herring - actually bytes was not used with the protocol crate!

Supplants #164 update prost
Fixes #221 update jni