SSv2: Add an "authentication tag" to verify the sender's identity
SSv1 included a "key agreement" between the sender and recipient in
the derivation of the message key, but SSv2 necessarily can't do that
(because the message is encrypted once for all recipients). Instead,
add an "authentication tag" derived from the key agreement, which can
be checked from the sender's identity included in the encrypted
message.