Android: force-link libdl
For some reason, the Docker build of libsignal_jni.so doesn't link to
libdl, even though the build on my local machine does. This doesn't
seem to be a problem on most phones, presumably because the libdl
functions are re-exported through libc. However, on a small subset of
Android phones this is not good enough and libsignal_jni.so fails to
load. For now, work around this by forcing the Android
cross-compilation linker to link to libdl (instead of doing it "as
needed").