signal-neon-futures: Poll a future once before queuing it
Previously, when using the JavaScript context to run a Rust future,
the future would be put into the event queue immediately, without
starting it first. The new implementation runs the future
synchronously to its first suspension point. There are pros and cons
to each approach (and I left the "delayed start" version behind as
`queue_future`), but in practice we're not worried about, say,
blocking other work behind this one.