signal-neon-futures: Only run Rust futures on an existing EventQueue
Node can more efficiently handle multiple tasks coming in on the same
queue, so remove the "convenience" APIs that derive a new queue from a
Context, and require an existing EventQueue instead. This cuts more
time off of our decryption benchmark (not checked in).

Additionally, run the first poll for the future synchronously, to
avoid having to wait for the event loop to pick up the task to start
the future.