A SenderKeyMessage's version must match the SenderKeyState
...or it should fail to decrypt. This parallels a similar check for
SignalMessage and SessionState. Since we haven't been storing this
field previously, treat a default "0" value as if it were version 3,
the current version.

...but wait, why is version 3 the current version? Historical
accident: we were reusing the version values from SignalMessages,
which currently only support version 3. This isn't unreasonable since
the SenderKey cipher is based on the usual session cipher, but that's
not enough reason to tie the two version numbers together in the
future. Stop doing that here.