Node: make sealedSenderDecryptToUsmc synchronous
...by requiring IdentityKeyStore.getIdentityKey and
getLocalRegistrationId to be synchronous.

Because of the way libsignal-client is built, this change propagates
to the other platforms' implementations, but the interface doesn't
change there.

In theory we could implement this at the Rust level by having a
"synchronous IdentityKeyStore trait" and the regular IdentityKeyStore
trait inherit from it, but that's a lot of trouble for the single
accessor sealed_sender_decrypt_to_usmc was using unconditionally.