Add DecryptionErrorMessage
This is a new message type that indicates that the current app was
unable to decrypt a received message (for instance, because they don't
have the sender key needed to decrypt a SenderKeyMessage). It includes
the timestamp of the original message as well as the "ratchet key" to
identify the 1:1 session it came from, if applicable.

This is a content message, like SenderKeyDistributionMessage. See next
commit.