Add a PlaintextContent, a CiphertextMessage for unencrypted content
For now this is just DecryptionErrorMessage, which is sent without
session encryption because something might be wrong with the session.
In the future we might have other kinds of PlaintextContent; what's
important is that we don't mix up content that's okay to send
unencrypted with content that should never be in the clear.

Note that even though the content might not be session-encrypted, it
can still be put in a sealed sender message, which has its own
encryption.