protocol: Deserialize archived sessions lazily instead of eagerly
The only operation that touches archived sessions is decryption, and
even then only in two cases: when checking to see if an incoming
PreKey message matches an existing session, and after decrypting on
the current session fails. Neither of those are unlikely, but they
happen an order of magnitude less than normal decryptions (not to
mention encryptions), so it's worth making the fast path faster.