Sealed Sender v2: avoid rederiving public keys unnecessarily
While a Curve25519 private key can always derive its corresponding
public key, doing so is not a free operation. Make sure we're only
ever doing so once, or not at all if we already have that information.
This results in a ~10% speedup to SSv2 encrypt and decrypt operations.

Also, convert the output of HKDF to a fixed-size array for storage,
instead of keeping around a boxed slice. This seems to allow for
further optimization and another 5% or so speedup.