Node: Test with plain 'mocha' instead of 'electron-mocha'
...which drops our dependency on Electron altogether. We originally
tested with electron-mocha to more closely resemble the Desktop app,
but libsignal-client doesn't actually use anything Electron-specific,
and because it uses N-API we don't have to sync up versions exactly
(and indeed we haven't been updating the Electron in this repo as
often as the Desktop app has taken new Electrons).

Two benefits of this: you can now run the tests on headless systems
(see the change to the CI script), and `yarn install` has less to
download.