attest: Remove 'packed' from declarations of SGX structs
Our use of `packed` matches the declaration of these structs as
`#pragma packed(1)` in C. However, 'packed' has unfortunate
implications, the biggest one being that it's unsafe to form
references to fields within these structs. Recent versions of Rust
have gotten more strict about that, so rather than put 'packed' on all
our structs, add an assertion that the total alignment of these
structs is 1. This basically checks what 'packed' would do for us
automatically, and we have existing assertions about the total size of
these types that make sure we're staying in sync with C.