bridge: Parameterize AsyncRuntime by the Future type it has to execute
This will come in handy when bringing the existing Node async function
support into the AsyncRuntime model: these async operations do *not*
require Send, because they are executed synchronously with JavaScript
operations on the Node microtask queue. This differs from the
"concurrent" async we plan to use with bridge_io, which *will* require
Send on its Futures.