CI: Pull the 32-bit Rust tests out to separate jobs
The Rust jobs are currently the long pole in the CI tests, but the
32-bit tests have to recompile everything except the proc-macros, so
they add the most time on top of the other steps. By separating them
to their own jobs (one for stable, one for nightly), we pay a little
extra in CI-minutes to save in turnaround time.