Consistently pad encrypted usernames to 128 bytes
...by introducing a padding field in the protobuf that ensures that
the encoded bytes are always between three and four AES blocks long.

This is a backwards-compatible change on the decoding side. For
encoding, certain very long usernames will now be rejected that
weren't previously, but these usernames would have to have
out-of-range discriminators (or not match the standard username format
at all).