SSv2: Add send support for the compact device list format
This takes advantage of the fact that multiple devices for the same
user will have the same identity key and therefore will use the same
per-recipient SSv2 data anyway.

This commit also enforces (on the client send side) that device IDs
are in the range 1..=127 for destinations of a SSv2 message.