Add validation for backup ChatItem frames
Check UUIDs and lengths of binary keys, and add some missing foreign key 
checks. The guiding principles for this approach are

a) using types that witness to correctness of the converted values
b) using a simple recursive structure for fallible conversion
c) exhaustively destructuring and matching to prevent missing checks

While this ends up producing verbose code, the code itself is simple and more 
likely to be obviously correct.